Excel protfolio project no 1000 ( Vrindavan store)
Hi Connection,

I feel so excited to share that I am going to set a challenge for myself I will share 1000 hands-on Projects related to Data Science and Data Analytics to showcase my work in Data Science.

And this Excel project is the 1000th project of that challenge and all the posts are going to be in Deceasing order. Like 1000,999,998,987….. Etc.
Well after that you will find some help from me.

I am feeling happy to share that this Excel Project is the first step of that journey. I hope you guys will like this challenge.

The objective of this Project:-  Create an interactive dashboard to show what's going on in the Vrindavan store like Order Vs Sales, Sales: Men Vs Women, Order Status, Top 5 States in terms of sales, Orders: Age Vs Gender and Sales via different channels.  

Compare the sales and orders using a single chart.
Which month got the high sales order ?
Who purchased more men or women in 2022.?
What are the different order statuses in 2022.?
list of top 10 states contributing to the sales.
Relation between age and gender based on the number of orders?
Which channel is contributing to maximum sales?














 
 
 
 

Steps taking by me while making this project:-

1):- Data cleaning        2):- Data Processing        3):- Dashboard

1)  Data cleaning:- 
    First thing first you must take data and start cleaning the checks column, Insert a new column, eliminate free space, etc

2) Data Processing:-
Insert column and apply text function to get a month in text.
Insert the column and apply the if function to get the age group.
And after that cross check the entire sheet to find the loop whole and after cross-checking make a pivot table according to demand like Sales: Men vs Women, etc
And then copy and paste all pivot charts on a single sheet and connect every silencer with each chart and then your dashboard is ready.
.

Insights provided:-

Women are more likely to buy compared to men.
Maharashtra, Karnataka and Uttar Pradesh are the top 3 states in sales.
The adult age group (30-49) is max contributing.
Amazon, Flipkart and Myntra channels are max contributing

Final Conclusion to improve the store sales:

Target Women customers of the age group (30-49) living in Maharashtra, Karnataka and Uttar Pradesh
 By running ads and offering coupons and discounts on Amazon, Flipkart and Myntra.

Special thanks to Risab Mishra from whom I got data of the Vrindavan Store.


 #excel #exceldashboards #sql #1000projects #underdog #darkhorse #portfolio project #powerbi #tableau #data #dataanalytics #dataanalysis #dataanalyst #datacleaning #dataanalystjobs #datacareers #dataprocessing #barclays #accenture #msexcel